Simplify update-authors.py
Before this change, there was effectively two for loops, one for the
list comprehension and one to iterate over the resulting list.
Generating the list wasn’t necessary since the code would just
immediately iterate over the list after it was generated, and then never
do anything with the list ever again.